Allow symlinks outside of mp3 folder
Allow symlinks outside the specified mp3 folders. If a symlink points to a directory that hasn't been scanned yet, scan it.
Because a directory can now be scanned (via a symlink) before "itself", we now check if a fid has already been given to that directory. If so, simply add a fid to that directory instead of rescanning it.